+++ ChangeLog.md
@@ -5,6 +5,11 @@
 
 # SW360 Base Library for Python
 
+## UNRELEASED
+
+* use SW360 support for `update_project_releases(..., add=True)` instead of our own code.
+  This allows to pass a dict with release details (e.g. project mainline state or relation).
+
 ## V1.8.1
 
 * `delete_project` now works properly. Before it could have happened the you get a `JSONDecodeError`.